pyproject.toml
[tool.poetry]
name = "djangoshop"
version = "0.1.0"
description = "Test project for a Django grocery shop"
authors = ["smintank <assbestmail@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.6"
python-dotenv = "^1.0.0"
pillow = "^10.3.0"
django-rest-framework = "^0.1.0"
djoser = "^2.2.3"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
django-rest-swagger = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
